1890 different cards.
Generating 90 unique cards per person, for 21 persons. To do so, I have
exported every page of my scribus document as an EPS so I keep the
generative aspect of it, even if I've been layout it graphically.

The *master* cards are then converted to PDF 90 times each, using
`epstopdf`. I then use PodofoImpose to recombine all of them. I have
prepared an A3 with the crop marks to superimpose on top of the imposed
cards, because Scribus doesn't allow for cropmarks with the EPS export.